Use percentage() to combine fuel corrections within a shared function Say what we mean! :-)

+++ speeduino/maths.h
@@ -198,7 +198,7 @@ static constexpr uint8_t ONE_HUNDRED_PCT = 100U;
  * @param value The value to operate on
  * @return uint32_t 
  */
-static inline uint32_t percentage(uint8_t percent, uint32_t value) 
+static inline uint32_t percentage(uint16_t percent, uint32_t value) 
 {
     return (uint32_t)div100((uint32_t)value * (uint32_t)percent);
 }
